backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
Pillow==9.3.0
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
celery==4.4.1
redis==3.4.1
PyYAML==6.0
